# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nasdaq option chain",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/research/runs/{run_id}": {
      "get": {
        "summary": "Run",
        "operationId": "run_api_research_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/runs/{run_id}/tickers": {
      "get": {
        "summary": "Run Tickers",
        "operationId": "run_tickers_api_research_runs__run_id__tickers_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunTickerView"
                  },
                  "title": "Response Run Tickers Api Research Runs  Run Id  Tickers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/overview": {
      "get": {
        "summary": "Overview",
        "operationId": "overview_api_research_overview_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OverviewCard"
                  },
                  "title": "Response Overview Api Research Overview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/leaderboard": {
      "get": {
        "summary": "Leaderboard",
        "operationId": "leaderboard_api_research_leaderboard_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ticker"
            }
          },
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "name": "min_trades",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Trades"
            }
          },
          {
            "name": "include_rejected",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Rejected"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "robustness",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "group_variants",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Group Variants"
            }
          },
          {
            "name": "rule",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule"
            }
          },
          {
            "name": "exit_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exit Kind"
            }
          },
          {
            "name": "filters",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filters"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Leaderboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/cross-ticker": {
      "get": {
        "summary": "Cross Ticker",
        "operationId": "cross_ticker_api_research_cross_ticker_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrossTickerItem"
                  },
                  "title": "Response Cross Ticker Api Research Cross Ticker Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/strategies/{strategy_id}": {
      "get": {
        "summary": "Strategy Detail",
        "operationId": "strategy_detail_api_research_strategies__strategy_id__get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          },
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9][A-Z0-9.\\-]{0,11}$",
              "title": "Ticker"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CrossTickerItem": {
        "properties": {
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "signals": {
            "type": "string",
            "title": "Signals"
          },
          "parameters": {
            "type": "string",
            "title": "Parameters"
          },
          "cross_score": {
            "type": "number",
            "title": "Cross Score"
          },
          "per_ticker": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CrossTickerPeer"
            },
            "type": "object",
            "title": "Per Ticker"
          },
          "entry_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entry Signals"
          },
          "filter_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filter Signals"
          },
          "exit_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exit Signals"
          },
          "exit_kind": {
            "type": "string",
            "title": "Exit Kind",
            "default": "mirror"
          }
        },
        "type": "object",
        "required": [
          "strategy_id",
          "strategy",
          "family",
          "signals",
          "parameters",
          "cross_score",
          "per_ticker"
        ],
        "title": "CrossTickerItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Leaderboard": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "families": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Families"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "total",
          "families",
          "items"
        ],
        "title": "Leaderboard"
      },
...
      "OverviewCard": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "bars": {
            "type": "integer",
            "title": "Bars"
          },
          "first": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First"
          },
          "last": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last"
          },
          "limited_history": {
            "type": "boolean",
            "title": "Limited History"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Id"
          },
          "strategy_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Name"
          },
          "signals": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signals"
          },
          "robustness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Robustness"
          },
          "oos_cagr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oos Cagr"
          },
          "sharpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe"
          },
          "max_drawdown": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown"
          },
          "trades": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trades"
          },
          "buy_hold_cagr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Hold Cagr"
          },
          "buy_hold_max_drawdown": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Hold Max Drawdown"
          },
          "test_cagr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Cagr"
          },
          "test_max_drawdown": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Max Drawdown"
          },
          "test_trades": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Trades"
          },
          "test_buy_hold_cagr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Buy Hold Cagr"
          },
          "entry_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entry Signals"
          },
          "filter_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filter Signals"
          },
          "exit_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exit Signals"
          },
          "exit_kind": {
            "type": "string",
            "title": "Exit Kind",
            "default": "mirror"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "bars",
          "first",
          "last",
          "limited_history"
        ],
        "title": "OverviewCard"
      },
...
      "RunTickerView": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "n_bars": {
            "type": "integer",
            "title": "N Bars"
          },
          "first_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Ts"
          },
          "last_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ts"
          },
          "limited_history": {
            "type": "boolean",
            "title": "Limited History"
          },
          "survivors": {
            "type": "integer",
            "title": "Survivors"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "n_bars",
          "limited_history",
          "survivors"
        ],
        "title": "RunTickerView"
      },
      "RunView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "strategy_count": {
            "type": "integer",
            "title": "Strategy Count"
          },
          "ticker_count": {
            "type": "integer",
            "title": "Ticker Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "status",
          "strategy_count",
          "ticker_count"
        ],
        "title": "RunView"
      },
...
      "StrategyDetail": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "signals": {
            "type": "string",
            "title": "Signals"
          },
          "entry_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entry Signals"
          },
          "filter_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filter Signals"
          },
          "exit_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exit Signals"
          },
          "exit_kind": {
            "type": "string",
            "title": "Exit Kind",
            "default": "mirror"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "type": "object",
            "title": "Parameters"
          },
          "entry": {
            "$ref": "#/components/schemas/ConditionGroup"
          },
          "exit": {
            "$ref": "#/components/schemas/ConditionGroup"
          },
          "robustness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Robustness"
          },
          "degradation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degradation"
          },
          "stability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stability"
          },
          "walk_forward_consistency": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Walk Forward Consistency"
          },
          "rejected": {
            "type": "boolean",
            "title": "Rejected"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "data_snapshot": {
            "type": "string",
            "enum": [
              "exact",
              "changed"
            ],
            "title": "Data Snapshot",
            "default": "exact"
          },
          "segment_bounds": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SegmentBound"
            },
            "type": "object",
            "title": "Segment Bounds"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentView"
            },
            "type": "array",
            "title": "Segments"
          },
          "folds": {
            "items": {
              "$ref": "#/components/schemas/FoldView"
            },
            "type": "array",
            "title": "Folds"
          },
          "reopt": {
            "items": {
              "$ref": "#/components/schemas/ReoptView"
            },
            "type": "array",
            "title": "Reopt"
          },
          "dates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dates"
          },
          "close": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Close"
          },
          "entry_marks": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Entry Marks"
          },
          "exit_marks": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Exit Marks"
          },
          "equity": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Equity"
          },
          "buy_hold": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Buy Hold"
          },
          "drawdown": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Drawdown"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeView"
            },
            "type": "array",
            "title": "Trades"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "ticker",
          "strategy_id",
          "name",
          "family",
          "signals",
          "parameters",
          "entry",
          "exit",
          "rejected",
          "flags",
          "segments",
          "folds",
          "reopt",
          "dates",
          "close",
          "entry_marks",
          "exit_marks",
          "equity",
          "buy_hold",
          "drawdown",
          "trades"
        ],
        "title": "StrategyDetail"
      },
...
    }
  }
}